Algorithms

Algorithms

Algorithms module includes a set of widely used MAb algorithms as weel as a common API to access them.

All the algorithms must be a subtype of MAB.Algorithms.BanditAlgorithmBase. See List of Agents for the list of available agents.

MAB.Algorithms.BanditAlgorithmBase — Type.

BanditAlgorithmBase

BanditAlgorithmBase defines a set of functions as well as some utility functions for implementing MAB algorithms. Any algorithm must be a subtype of BanditAlgorithmBase.

source

High Level interface

The folowing methods are defined for all MAB agents.

MAB.Algorithms.get_arm_index — Function.

get_arm_index( agent::BanditAlgorithmBase )

Gets the index of next arm to pull.

Example

source

MAB.Algorithms.update_reward! — Function.

update_reward!( agent::BanditAlgorithmBase, r::Real )

Updates the reward to bandit algorithm agent. r must a real number within valid range.

source

MAB.Algorithms.reset! — Function.

reset!( agent::BanditAlgorithmBase )

Resets the internal statistics of the bandit algorithm, except for the number of arms and other algorithm specific parameters. reset!() will make agent as fresh as it was first created.

source

MAB.Algorithms.info_str — Function.

info_str( agent::BanditAlgorithmBase, latex::Bool = false )

Return a information string about the agent and it's parameters. If flag latex is set to true, then the returned string will be a compactable latex string.

source

Utility Functions

Further, Algorithms module provides the following utility functions, but are not exported implicitly. They must be referred in their full scope on call or must be explicitly imported to user scripts.

MAB.Algorithms.make_agents_with_k — Function.

make_agents_with_k( K::Int64, agent_list::Vector{} )

Returns a vector of agents specified in agent_list with K arms.

source

Base.:== — Function.

==

Can be used to compare two agents.

source