Home

MAB.jl

A Julia Package for Experimenting with Multi Arm Bandits

MAB — Module.

MAB.jl is a package for experimenting with multi arm bandit algorithms. This package comes with different MAB algorithms, various arm models and a module to reproduce/evaluate your own algorithms in different scenarios/experiments which are already available in literature.

This package is divied into three main modules

Algorithms

ArmModels

Experiments

source

Manual

Package Guide

Installing MAB.jl package

Usage

Accessing documentation from REPL

Package Outline

Common API

Algorithms

Arm Models

Experiments